tools/buildtools/dev_requirements.txt
mock==1.0.1
nose==1.3.4
vcrpy==1.7.4
flake8
